PHP Talent Test
Your company would like to develop an online dice game. The game rules
are as below:

1. The game contains 4 players.
2. Each player will have 6 dice in their dice up.
3. Each round all players will roll their dice at the same time.
4. All dice with number 1 on top side will be passed to player on his
right hand (the right most player will pass the dice to left most
player)
5. All dice with number 6 on top side will be removed from their dice
cup.
6. All players roll their dice again to start next round.
7. The player who first emptied their dice cup (could be more than 1
player) is winner(s).

You are required to write a PHP code to simulate this game and print out
the result of each round until winner(s) is found.